requirements.txt
astunparse==1.6.3
certifi==2022.12.7
ftfy==6.1.1
jclip==1.0
jittor==1.3.9.4
joblib==1.3.2
numpy==1.21.6
Pillow==9.5.0
pip==22.3.1
regex==2023.12.25
scikit-learn==1.0.2
scipy==1.7.3
setuptools==65.6.3
six==1.16.0
threadpoolctl==3.1.0
tqdm==4.66.2
wcwidth==0.2.13
wheel==0.38.4
